Hw2

+++ link-parser/src/main/java/ru/tinkoff/edu/java/linkParser/records/StackOverflowInfo.java
@@ -1,6 +1,6 @@
 package ru.tinkoff.edu.java.linkParser.records;
 
-public record StackOverflowInfo(int id) implements LinkInfo {
+public record StackOverflowInfo(long id) implements LinkInfo {
     public String toString() {
         return String.valueOf(id);
     }
